feat(extension): move Bandcamp feed sync into the background worker
The popup's Feed button was disabled on every *.bandcamp.com artist
subdomain because the content-script path issued credentialed POSTs
against bandcamp.com and SameSite handling on the session cookie made
those requests unreliable from a subdomain origin. Move the scrape
into the service worker, which runs at the extension's own origin and
already has the *.bandcamp.com host-permission grant — credentialed
fetches succeed regardless of which Bandcamp tab the user has open.

Worker side:
- Add scrapeFeedFromWorker({ pageCount }) in service_worker.js that
  reuses the parseFeedPage / assertJsonContentType helpers shipped
  in the bandcamp-feed-defensive-parse change. Releases flow through
  a new ingestBandcampFeedReleases({ data, done }) helper that
  directly mutates the worker's release accumulator and triggers
  fetchNextBandcampItem on the final page.
- Add a bandcamp:scrape-feed handler that delegates to the new
  helper, routing failures through handleError so they reach the
  popup error UI on the same channel as every other worker error.

Popup side:
- BandcampPanel.sendFeed now calls
  browser.runtime.sendMessage({ type: 'bandcamp:scrape-feed' })
  directly. The Feed button drops the onSubdomain gate; only
  `running || !loggedIn` remain.

Content-script side:
- Replace the bandcamp:scrape-feed case with a one-line forwarder so
  any tabs.sendMessage caller still using the legacy path falls
  through to the worker.
- Delete the now-unused scrapeFeed and reportProgress helpers and the
  parse-helper imports — they live in feed-parse.js, imported only by
  the worker now.

Backlog also adds item 025 (include "New Releases from artists you
follow" panel in the feed sync), at top priority — that's the next
gap once subdomain-blocking is resolved.

Verification: yarn test green (13 specs); FRONTEND_URL=...
yarn build:chrome compiles cleanly; manifest.base.json byte-identical
to HEAD.

Closes backlog item 021.
