+++ backlog/done/021-bandcamp-feed-sync-from-subdomain
