+++ backlog/tasks/025-bandcamp-feed-include-followed-artist-releases/notes.md
@@ -1,0 +1,20 @@
+# Notes
+
+Working notebook for this item. Date entries so future sessions can skim.
+
+## Decisions
+
+- _YYYY-MM-DD_ — decision and reasoning
+
+## Rejected approaches
+
+- _YYYY-MM-DD_ — what was tried, why it didn't work. Save the next session from
+  retrying it.
+
+## Open threads
+
+- Things to come back to, partially-explored ideas, suspicious findings
+
+## Session log
+
+- _YYYY-MM-DD_ — short note on what was done / discovered this session
